refactor: patch requests -> patchsets -> patches
Previously a patch request contained a series of patches.  This worked
great as an MVP but we are starting to see some issues with this impl.

Previously the contrib and reviewer would push patches to git-pr similar
to github pull requests.  They all get merged into a single patchset.
This mostly works until you want to start editing previous commits to
keep the commit history tidy and relevant.  For many workflows, going
back to a previous commit and amending it to address feedback is
desirable.

This creates a new model, patchset, which is a mostly immutable
container for patches.  1-to-many patch request to patchsets and
1-to-many patchset to patches.  Think of these patchsets as revisions.

This allows us to better organize collaboration and enable features like
`git range-diff` to see changes between revisions.

BREAKING CHANGE: sqlite dbs will have to be recreated as the new models
are fundamentally different. Sorry for the inconvenience!