feat: empty patch support
`git format-patch --cover-letter` is a great way to provide a summary
of all the changes within a patchset so we want to explicity support
that functionality.

The only gotcha here is we now allow empty patches.  So if there is an
empty patch in the patchset, the committer needs to add another option:
`git am -3 --empty=stop|drop|keep` depending on their preference.