feat: add multitenancy
The goal of this change is to support the ability to let any user create
repos, similar to popular code forges.  Repos are created adhoc when a
user uploads a patchset.

As part of this change, we are deprecating the previous way to configure
repos.  Repos can be thought of as containers for patch requests and not
much more.  They don't need to strictly map to a git repo.

Multitenancy is opt-in via the `git-pr.toml` field:

```
create_repo = "user"
```